move crypto,boost,parser to thirdparty
git-svn-id: http://192.168.20.11/sequoiadb/trunk@11622 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dms/dmsStorageData.cpp
@@ -1005,7 +1005,7 @@ namespace engine
    void _dmsStorageData::_mapExtent2DelList( dmsMB * mb, dmsExtent * extAddr,
                                              SINT32 extentID )
    {
-      if ( extAddr->_freeSpace < DMS_MIN_RECORD_SZ )
+      if ( (UINT32)extAddr->_freeSpace < DMS_MIN_RECORD_SZ )
       {
          if ( extAddr->_freeSpace != 0 )
          {
